refactor(markdown): extract @kolu/solid-markdown from the intent surface
Lift the marked-based token-walk renderer out of
client/src/intent/IntentMarkdown.tsx (257→38 LOC) into a standalone,
app-agnostic package. One <Markdown> component with inline/compact/
document variants bundling parse mode + styling scale; the intent
surface now consumes it via thin wrappers that pin variant + link
policy, so the ~8 call sites stay untouched (behaviour-preserving).

The "compact" variant reproduces the intent styling byte-for-byte;
"document" is the reading-scale variant the planned Code-tab Source ⇄
Rendered preview will consume (Phase 1 of the solid-fileview plan).

An @source directive in index.css keeps Tailwind v4 generating the
package's utility classes (verified: the .67rem/.85em/.8rem classes —
unique to this package — appear in the built CSS).
