+++ packages/client/src/index.css
@@ -1,6 +1,12 @@
 @import "kolu-fonts";
 @import "tailwindcss";
 
+/* `@kolu/solid-markdown` ships its own Tailwind utility classes (the
+ * markdown renderer's styling variants). Make the cross-package source
+ * explicit so those classes are always generated — Tailwind v4's
+ * auto-detection base shouldn't be relied on to reach a sibling package. */
+@source "../../solid-markdown/src";
+
 /* Suppress the browser's pull-to-refresh and scroll-chaining at the
  * document level. Without this, swiping down at the top of the
  * sidebar or terminal viewport hijacks the gesture into a page
