+++ packages/solid-markdown/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "@kolu/solid-markdown",
+  "version": "0.1.0",
+  "private": true,
+  "license": "AGPL-3.0-or-later",
+  "type": "module",
+  "sideEffects": false,
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "scripts": {
+    "typecheck": "tsc --noEmit"
+  },
+  "devDependencies": {
+    "typescript": "^5.8.0"
+  },
+  "dependencies": {
+    "marked": "^18.0.2",
+    "solid-js": "^1.9.0"
+  }
+}
