+++ packages/solid-markdown/src/index.ts
@@ -1,0 +1,1 @@
+export { Markdown, type MarkdownVariant } from "./Markdown";
